# {{cookiecutter.project_slug}}/docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "企业级FastAPI后端模板",
    "description": "企业级FastAPI后端模板 - 开箱即用",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/role/list": {
      "get": {
        "tags": [
          "角色模块"
        ],
        "summary": "查看角色列表",
        "operationId": "list_role_api_v1_role_list_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "页码",
              "default": 1,
              "title": "Page"
            },
            "description": "页码"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "每页数量",
              "default": 10,
              "title": "Page Size"
            },
            "description": "每页数量"
          },
          {
            "name": "role_name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "角色名称，用于查询",
              "default": "",
              "title": "Role Name"
            },
            "description": "角色名称，用于查询"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/role/get": {
      "get": {
        "tags": [
          "角色模块"
        ],
        "summary": "查看角色",
        "operationId": "get_role_api_v1_role_get_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "role_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "角色ID",
              "title": "Role Id"
            },
            "description": "角色ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/role/create": {
      "post": {
        "tags": [
          "角色模块"
        ],
        "summary": "创建角色",
        "operationId": "create_role_api_v1_role_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/role/update": {
      "post": {
        "tags": [
          "角色模块"
        ],
        "summary": "更新角色",
        "operationId": "update_role_api_v1_role_update_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/role/delete": {
      "delete": {
        "tags": [
          "角色模块"
        ],
        "summary": "删除角色",
        "operationId": "delete_role_api_v1_role_delete_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "role_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "角色ID",
              "title": "Role Id"
            },
            "description": "角色ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/role/authorized": {
      "get": {
        "tags": [
          "角色模块"
        ],
        "summary": "查看角色权限",
        "operationId": "get_role_authorized_api_v1_role_authorized_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "角色ID",
              "title": "Id"
            },
            "description": "角色ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "角色模块"
        ],
        "summary": "更新角色权限",
        "operationId": "update_role_authorized_api_v1_role_authorized_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleUpdateMenusApis"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/api/list": {
      "get": {
        "tags": [
          "API模块"
        ],
        "summary": "查看API列表",
        "operationId": "list_api_api_v1_api_list_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "页码",
              "default": 1,
              "title": "Page"
            },
            "description": "页码"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "每页数量",
              "default": 10,
              "title": "Page Size"
            },
            "description": "每页数量"
          },
          {
            "name": "path",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "API路径",
              "title": "Path"
            },
            "description": "API路径"
          },
          {
            "name": "summary",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "API简介",
              "title": "Summary"
            },
            "description": "API简介"
          },
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "API模块",
              "title": "Tags"
            },
            "description": "API模块"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/api/get": {
      "get": {
        "tags": [
          "API模块"
        ],
        "summary": "查看Api",
        "operationId": "get_api_api_v1_api_get_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Api",
              "title": "Id"
            },
            "description": "Api"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/api/create": {
      "post": {
        "tags": [
          "API模块"
        ],
        "summary": "创建Api",
        "operationId": "create_api_api_v1_api_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/api/update": {
      "post": {
        "tags": [
          "API模块"
        ],
        "summary": "更新Api",
        "operationId": "update_api_api_v1_api_update_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/api/delete": {
      "delete": {
        "tags": [
          "API模块"
        ],
        "summary": "删除Api",
        "operationId": "delete_api_api_v1_api_delete_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "api_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "ApiID",
              "title": "Api Id"
            },
            "description": "ApiID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/api/refresh": {
      "post": {
        "tags": [
          "API模块"
        ],
        "summary": "刷新API列表",
        "operationId": "refresh_api_api_v1_api_refresh_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiCreate": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path",
            "description": "API路径",
            "example": "/api/v1/user/list"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "description": "API简介",
            "default": "",
            "example": "查看用户列表"
          },
          "method": {
            "$ref": "#/components/schemas/MethodType",
            "description": "API方法",
            "example": "GET"
          },
          "tags": {
            "type": "string",
            "title": "Tags",
            "description": "API标签",
            "example": "User"
          }
        },
        "type": "object",
        "required": [
          "path",
          "method",
          "tags"
        ],
        "title": "ApiCreate"
      },
      "ApiUpdate": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path",
            "description": "API路径",
            "example": "/api/v1/user/list"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "description": "API简介",
            "default": "",
            "example": "查看用户列表"
          },
          "method": {
            "$ref": "#/components/schemas/MethodType",
            "description": "API方法",
            "example": "GET"
          },
          "tags": {
            "type": "string",
            "title": "Tags",
            "description": "API标签",
            "example": "User"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "path",
          "method",
          "tags",
          "id"
        ],
        "title": "ApiUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MethodType": {
        "type": "string",
        "enum": [
          "GET",
          "POST",
          "PUT",
          "DELETE",
          "PATCH"
        ],
        "title": "MethodType"
      },
...
      "RoleCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "example": "管理员"
          },
          "desc": {
            "type": "string",
            "title": "Desc",
            "default": "",
            "example": "管理员角色"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "RoleCreate"
      },
      "RoleUpdate": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "example": 1
          },
          "name": {
            "type": "string",
            "title": "Name",
            "example": "管理员"
          },
          "desc": {
            "type": "string",
            "title": "Desc",
            "default": "",
            "example": "管理员角色"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "RoleUpdate"
      },
      "RoleUpdateMenusApis": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "menu_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Menu Ids",
            "default": []
          },
          "api_infos": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Api Infos",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "RoleUpdateMenusApis"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}